add KdtreePerfTest
